requirements.txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
branca==0.8.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cramjam==2.11.0
cycler==0.12.1
fastparquet==2024.11.0
filelock==3.20.0
folium==0.20.0
fonttools==4.60.1
fsspec==2025.10.0
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
googlemaps==4.10.0
hf-xet==1.2.0
httplib2==0.31.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
narwhals==2.12.0
networkx==3.6
numba==0.62.1
numpy==2.3.5
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.0.0
plotly==6.5.0
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.1.2
shap==0.50.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
statsmodels==0.14.5
streamlit==1.51.0
streamlit-folium==0.25.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
xgboost==3.1.2
xyzservices==2025.10.0
